# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.AddClazzMemberRequest:
    properties:
      member_id:
        type: string
    required:
    - member_id
    type: object
  models.AddTaskRequest:
    properties:
      clazz_id:
        type: string
      course_id:
        type: string
      description:
        type: string
      end_time:
        description: 如果不设置则没有结束时间
        type: string
      relation_ids:
        items:
          type: string
        type: array
      start_time:
        description: 必须设置开始时间
        type: string
      title:
        type: string
      type:
        $ref: '#/definitions/models.TaskType'
    required:
    - clazz_id
    - course_id
    - start_time
    - title
    - type
    type: object
...
  models.CourseStatus:
    enum:
    - 0
    - 1
    - 2
    format: int32
    type: integer
    x-enum-comments:
      CourseStatusActive: 活跃
      CourseStatusArchived: 已归档
      CourseStatusDraft: 草稿
    x-enum-descriptions:
    - 草稿
    - 活跃
    - 已归档
    x-enum-varnames:
    - CourseStatusDraft
    - CourseStatusActive
    - CourseStatusArchived
  models.CreateClazzRequest:
    properties:
      course_id:
        type: string
      description:
        type: string
      max_members:
        type: integer
      name:
        type: string
      require_invite:
        type: boolean
      schedule:
        type: string
    required:
    - course_id
    - name
    - require_invite
    type: object
  models.CreateCourseRequest:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - name
    type: object
...
  models.FinishTaskRequest:
    properties:
      clazz_id:
        type: string
      relation_id:
        type: string
      task_id:
        type: string
    required:
    - clazz_id
    - relation_id
    - task_id
    type: object
...
  models.PageQueryCourseRequest:
    properties:
      name:
        type: string
      page_num:
        type: integer
      page_size:
        type: integer
      status:
        type: integer
    type: object
...
  models.RemoveClazzMembersRequest:
    properties:
      member_ids:
        items:
          type: string
        type: array
    required:
    - member_ids
    type: object
...
  models.TaskType:
    enum:
    - 1
    - 2
    format: int32
    type: integer
    x-enum-comments:
      TaskTypeProblem: 题单
      TaskTypeVideo: 视频
    x-enum-descriptions:
    - 题单
    - 视频
    x-enum-varnames:
    - TaskTypeProblem
    - TaskTypeVideo
...
  models.UpdateCourseRequest:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
      status:
        $ref: '#/definitions/models.CourseStatus'
    required:
    - id
    type: object
  models.UpdateTaskRequest:
    properties:
      clazz_id:
        type: string
      course_id:
        type: string
      description:
        type: string
      end_time:
        type: string
      relation_ids:
        items:
          type: string
        type: array
      start_time:
        type: string
      task_id:
        type: string
      title:
        type: string
      type:
        $ref: '#/definitions/models.TaskType'
    required:
    - clazz_id
    - course_id
    - task_id
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@zkcodearena.com
    name: API Support
    url: http://localhost/support
  description: |-
    ZK Code Arena 在线编程平台 REST API 文档
    提供用户管理、题库管理、班级课程管理、代码评测等功能
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://localhost/terms/
  title: ZK Code Arena API
  version: "1.0"
paths:
  /courses:
    post:
      consumes:
      - application/json
      description: 教师创建新课程
      parameters:
      - description: 课程信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateCourseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功，返回课程ID
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 创建课程（教师）
      tags:
      - 课程
    put:
      consumes:
      - application/json
      description: 更新课程的基本信息
      parameters:
      - description: 更新的课程信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateCourseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 更新课程信息
      tags:
      - 课程
  /courses/{courseId}:
    delete:
      consumes:
      - application/json
      description: 删除指定课程
      parameters:
      - description: 课程ID
        in: path
        name: courseId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 删除课程
      tags:
      - 课程
    get:
      consumes:
      - application/json
      description: 根据课程ID获取课程详细信息
      parameters:
      - description: 课程ID
        in: path
        name: courseId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 课程详情
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取课程详情
      tags:
      - 课程
  /courses/{courseId}/{clazzId}:
    put:
      consumes:
      - application/json
      description: 刷新班级的邀请二维码
      parameters:
      - description: 课程ID
        in: path
        name: courseId
        required: true
        type: string
      - description: 班级ID
        in: path
        name: clazzId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 刷新成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 刷新班级二维码
      tags:
      - 课程
  /courses/{courseId}/avatar:
    put:
      consumes:
      - multipart/form-data
      description: 上传并更新课程头像图片
      parameters:
      - description: 课程ID
        in: path
        name: courseId
        required: true
        type: string
      - description: 头像文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 更新课程头像
      tags:
      - 课程
  /courses/clazzes:
    post:
      consumes:
      - application/json
      description: 为指定课程创建新班级
      parameters:
      - description: 班级信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateClazzRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 为课程创建班级
      tags:
      - 课程
  /courses/clazzes/{clazzId}:
    delete:
      consumes:
      - application/json
      description: 删除指定班级
      parameters:
      - description: 班级ID
        in: path
        name: clazzId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 删除班级
      tags:
      - 课程
    get:
      consumes:
      - application/json
      description: 根据班级ID获取班级详细信息
      parameters:
      - description: 班级ID
        in: path
        name: clazzId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 班级详情
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取班级详情
      tags:
      - 课程
  /courses/clazzes/{clazzId}/members:
    post:
      consumes:
      - application/json
      description: 手动添加成员到班级
      parameters:
      - description: 班级ID
        in: path
        name: clazzId
        required: true
        type: string
      - description: 成员ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.AddClazzMemberRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 添加成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 添加班级成员
      tags:
      - 课程
  /courses/clazzes/{clazzId}/members/remove:
    post:
      consumes:
      - application/json
      description: 批量移除班级成员
      parameters:
      - description: 班级ID
        in: path
        name: clazzId
        required: true
        type: string
      - description: 成员ID列表
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RemoveClazzMembersRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 移除成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 移除班级成员
      tags:
      - 课程
  /courses/clazzes/join:
    get:
      consumes:
      - application/json
      description: 学生通过邀请码加入班级
      parameters:
      - description: 班级ID
        in: query
        name: clazzId
        required: true
        type: string
      - description: 邀请码
        in: query
        name: invite_code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 加入成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 加入班级
      tags:
      - 课程
  /courses/finishtask:
    post:
      consumes:
      - application/json
      description: 标记任务为已完成
      parameters:
      - description: 完成任务信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.FinishTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 完成成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 完成任务
      tags:
      - 课程
  /courses/query:
    post:
      consumes:
      - application/json
      description: 分页查询用户相关的课程列表
      parameters:
      - description: 分页查询参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.PageQueryCourseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 课程列表
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 分页查询课程
      tags:
      - 课程
  /courses/task:
    post:
      consumes:
      - application/json
      description: 为班级添加新任务
      parameters:
      - description: 任务信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.AddTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 添加成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 添加任务
      tags:
      - 课程
    put:
      consumes:
      - application/json
      description: 更新任务信息
      parameters:
      - description: 更新的任务信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 更新任务
      tags:
      - 课程
  /courses/task/{clazzId}:
    get:
      consumes:
      - application/json
      description: 根据班级ID获取任务列表
      parameters:
      - description: 班级ID
        in: path
        name: clazzId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 任务列表
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取班级任务列表
      tags:
      - 课程
  /courses/task/{taskId}:
    delete:
      consumes:
      - application/json
      description: 删除指定任务
      parameters:
      - description: 任务ID
        in: path
        name: taskId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 删除任务
      tags:
      - 课程
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"